+++ CHANGES.rst
@@ -12,6 +12,8 @@ Unreleased
 -   Fix response_wrapper type check in test client. :issue:`2831`
 -   Make the return type of ``MultiPartParser.parse`` more
     precise. :issue:`2840`
+-   Raise an error if converter arguments cannot be
+    parsed. :issue:`2822`
 
 Version 3.0.1
 -------------